+++ pojo/src/main/java/com/stanzaliving/website/response/dto/ApartmentShortResponseDto.java
@@ -25,7 +25,7 @@ public class ApartmentShortResponseDto implements Serializable {
 	
 	private String propertyCode;
 	
-	private Long societyId;
+	private Integer societyId;
 	
 	private Integer roomCount;
